Request parameters are determined as required

+++ Presentation/SimpleLibrary.WebAPI/Controllers/BookTypeController.cs
@@ -1,5 +1,7 @@
+using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using SimpleLibrary.Application;
 using SimpleLibrary.Core.Dtos;
 
@@ -16,15 +18,21 @@ namespace SimpleLibrary.WebAPI.Controllers
         }
         
         [HttpPost("add-booktype")]
-        public async Task<IActionResult> CreateBookType(BookTypeModelDto model)
+        public async Task<IActionResult> CreateBookType([FromBody,Required]BookTypeModelDto model)
         {
+            if (ModelState.IsValid is false)
+                return BadRequest();
+            
             var result = await _bookTypeService.CreateBookType(model);
             return this.Ok(result);
         }
         
         [HttpDelete("delete-booktype")]
-        public async Task<IActionResult> DeleteBookType(string type)
+        public async Task<IActionResult> DeleteBookType([FromQuery,Required]string type)
         {
+            if (ModelState.IsValid is false)
+                return BadRequest();
+            
             var result = await _bookTypeService.DeleteBookType(type);
             if (result is null)
                 return this.Ok(
@@ -37,8 +45,11 @@ namespace SimpleLibrary.WebAPI.Controllers
         }
         
         [HttpGet("get-books")]
-        public async Task<IActionResult> GetBooks([FromQuery] string type,[FromQuery]int currentPage)
+        public async Task<IActionResult> GetBooks([FromQuery,Required] string type,[FromQuery,Required]int currentPage)
         {
+            if (ModelState.IsValid is false)
+                return BadRequest();
+            
             var result =await _bookTypeService.GetBooksByBookType(type, currentPage);
             if (result is null)
             {
@@ -48,6 +59,7 @@ namespace SimpleLibrary.WebAPI.Controllers
                     Message=$"There is no BookType as {type}"
                 });
             } 
+            
             return this.Ok(result);
         }
     }
